Renamed the pipeline modules.
