+++ crates/bld_runner/src/pipeline/mod.rs
@@ -1,4 +1,4 @@
 pub mod traits;
-pub mod version1;
-pub mod version2;
+pub mod v1;
+pub mod v2;
 pub mod versioned;
